Updated whisper-engine to use fp16 on WebGPU and adjusted seeder to include fp16 ONNX files, replacing streaming upload with buffered approach for reliability. Expanded model file list to include encoder/decoder fp16 variants and updated seeding logic to upload fp16 content.

X-Lovable-Edit-ID: edt-451798ae-4fde-4600-8f82-579d76cef437

+++ src/modules/uns/core/hologram/whisper-engine.ts
@@ -182,9 +182,10 @@ export class WhisperEngine {
         console.log("[Whisper] vGPU init failed → WASM fallback:", err);
       }
 
-      // Use fp32 on WebGPU for maximum accuracy, q8 (quantized) on WASM for speed.
+      // fp16 on WebGPU: best accuracy/size tradeoff (59MB decoder vs 119MB fp32).
+      // q8 on WASM: smallest, fast, good enough for speech.
       // Both variants are pre-seeded in our storage bucket.
-      const dtype = this._device === "webgpu" ? "fp32" : "q8";
+      const dtype = this._device === "webgpu" ? "fp16" : "q8";
 
       this._onProgress?.({ status: "downloading", progress: 0 });
